integration/gitops.py
from git import Repo
import os

class GitOpsBot:
    def __init__(self, repo_path: str = "."):
        try:
            self.repo = Repo(repo_path)
        except:
            self.repo = None # Handling case where not a git repo for demo
        
    def process_updates(self, updates: list):
        """
        Updates: List of dicts {key, old, new, confidence}
        """
        high_confidence = [u for u in updates if u['confidence'] >= 0.95]
        low_confidence = [u for u in updates if u['confidence'] < 0.95]
        
        if high_confidence:
            self._handle_auto_commit(high_confidence)
            
        if low_confidence:
            self._handle_pr_creation(low_confidence)

    def _handle_auto_commit(self, updates):
        print(f"[GitOps] Auto-committing {len(updates)} updates.")
        # Logic: checkout branch, edit file, add, commit, push
        # For prototype, just print
        for u in updates:
            print(f"  - Auto-patched: {u['key']} -> {u['new']}")

    def _handle_pr_creation(self, updates):
        print(f"[GitOps] Creating PR for {len(updates)} updates.")
        for u in updates:
            print(f"  - REVIEW REQUIRED: {u['key']} -> {u['new']} (Conf: {u['confidence']})")
