requirements.txt
playwright>=1.40.0
apted>=1.0.3
transformers>=4.36.0
torch>=2.1.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.9
chromadb>=0.4.22
gitpython>=3.1.40
pydantic>=2.5.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
numpy>=1.26.0
